moving time functions to util.c
