+++ lib/util.c
@@ -341,3 +341,4 @@ int set_cpu(uint32_t core)
 }
 
 #endif
+
